bandwidth-test/requirements.txt
fastapi
uvicorn[standard]
speedtest-cli 
